fix: Foundry tests 59/59 passing
- CommitmentMerkleTree: fix uint32 overflow in TreeFull check (uint32(1)<<32==0)
- Vault.t.sol: fix vault address prediction nonce offset (+2 not +3)
- Vault.t.sol: fix invalid Solidity hex literals in test address constants
- Vault.t.sol: fix setVerifier onlyOwner test consuming prank on view call
- Vault.t.sol: fix naCancellationCredit nullifierSpent test using wrong nullifier

+++ packages/contracts/src/CommitmentMerkleTree.sol
@@ -55,7 +55,7 @@ contract CommitmentMerkleTree {
     /// @notice Insert a commitment leaf and return the new root.
     function insert(bytes32 commitment) external onlyVault returns (bytes32 newRoot) {
         uint32 index = nextIndex;
-        if (index >= uint32(1) << TREE_DEPTH) revert TreeFull();
+        if (uint256(index) >= uint256(1) << TREE_DEPTH) revert TreeFull();
         nextIndex = index + 1;
 
         bytes32 current = commitment;
